Migrate legacy packagers page to Symfony controller and Twig template
