+++ app/Resources/views/packagers/index.html.twig
@@ -1,0 +1,33 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Packer{% endblock %}
+
+{% block body %}
+    <table class="pretty-table">
+        <tr>
+            <th><a href="{{ path('app_packagers_index', {'orderby': 'name', 'sort': sort}) }}">Name</a></th>
+            <th>E-Mail</th>
+            <th colspan="2"><a
+                        href="{{ path('app_packagers_index', {'orderby': 'packages', 'sort': sort}) }}">Pakete</a>
+            </th>
+            <th><a href="{{ path('app_packagers_index', {'orderby': 'lastbuilddate', 'sort': sort}) }}">Letzte
+                    Aktualisierung</a></th>
+        </tr>
+        {% for packager in packagers %}
+            <tr>
+                <td>{{ packager.name }}</td>
+                <td>{% if packager.email %}<a href="mailto:{{ packager.email }}">{{ packager.email }}</a>{% endif %}
+                </td>
+                <td style="text-align:right;"><a
+                            href="{{ path('app_legacy_page', {'page': 'Packages', 'packager': packager.id}) }}">{{ packager.packages }}</a>
+                </td>
+                <td style="width:100px;">
+                    <div style="background-color:#1793d1;width:{{ 100 * packager.packages / packages | round }}px;">
+                        &nbsp;
+                    </div>
+                </td>
+                <td>{{ packager.lastbuilddate | date('d.m.Y') }}</td>
+            </tr>
+        {% endfor %}
+    </table>
+{% endblock %}
