+++ app/Resources/views/start/index.html.twig
@@ -25,7 +25,7 @@
     {% endfor %}
 {% endblock %}
 
-{% block title %}Start!{% endblock %}
+{% block title %}Start{% endblock %}
 
 {% block body %}
     <div id="left-wrapper">
